Changed Duration and Timestamp to long

+++ SpawnDev.BlazorJS/JSObjects/VideoFrameDataOptions.cs
@@ -24,12 +24,12 @@ namespace SpawnDev.BlazorJS.JSObjects
         /// <summary>
         /// An integer representing the timestamp of the frame in microseconds.
         /// </summary>
-        public double Timestamp { get; set; }
+        public long Timestamp { get; set; }
         /// <summary>
         /// An integer representing the duration of the frame in microseconds.
         /// </summary>
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-        public int? Duration { get; set; }
+        public long? Duration { get; set; }
         /// <summary>
         /// A list containing the layout for each plane
         /// </summary>
